server/requirements.txt
flask
google-cloud-pubsub
gunicorn
python-decouple
google-generativeai
beartype
google-cloud-firestore
